requirements.txt
anyio==4.6.0
asgiref==3.8.1
certifi==2024.8.30
Django==5.1.1
django-allauth==65.0.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
ollama==0.3.3
pillow==10.4.0
PyJWT==2.9.0
rest-framework-simplejwt==0.0.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
selenium==4.14.0
webdriver-manager==3.8.6
